remove nodejs-mobile-cordova
