--- config.xml
@@ -1,32 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<widget id="io.cordova.hellocordova" version="1.0.0" xmlns="http://www.w3.org/ns/widgets" xmlns:cdv="http://cordova.apache.org/ns/1.0">
-    <name>HelloCordova</name>
-    <description>
-        A sample Apache Cordova application that responds to the deviceready event.
-    </description>
-    <author email="dev@cordova.apache.org" href="http://cordova.io">
-        Apache Cordova Team
-    </author>
-    <content src="index.html" />
-    <access origin="*" />
-    <allow-intent href="http://*/*" />
-    <allow-intent href="https://*/*" />
-    <allow-intent href="tel:*" />
-    <allow-intent href="sms:*" />
-    <allow-intent href="mailto:*" />
-    <allow-intent href="geo:*" />
-    <platform name="android">
-        <allow-intent href="market:*" />
-        <resource-file src="local.properties" target="local.properties" />
-        <preference name="android-targetSdkVersion" value="29" />
-        <preference name="android-minSdkVersion" value="22" />
-        <preference name="AndroidXEnabled" value="true" />
-        <hook src="scripts/android/prepareNode.js" type="before_build" />
-        <hook src="scripts/android/removeZip.js" type="before_build" />
-        <hook src="scripts/android/zip.js" type="after_build" />
-    </platform>
-    <platform name="ios">
-        <allow-intent href="itms:*" />
-        <allow-intent href="itms-apps:*" />
-    </platform>
-</widget>
