+++ package.json
@@ -1,31 +1,31 @@
 {
-  "name": "io.cordova.hellocordova",
-  "displayName": "HelloCordova",
-  "version": "1.0.0",
-  "description": "A sample Apache Cordova application that responds to the deviceready event.",
-  "main": "index.js",
+  "version": "0.0.1",
+  "description": "RedMobile Custom Module Generator",
+  "main": "main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "bash prepare.sh && node main.js"
   },
-  "keywords": [
-    "ecosystem:cordova"
-  ],
-  "author": "Apache Cordova Team",
-  "license": "Apache-2.0",
   "dependencies": {
-    "archiver": "^5.2.0"
-  },
-  "devDependencies": {
-    "cordova-android": "^9.0.0",
-    "cordova-plugin-whitelist": "^1.3.4",
+    "@red-mobile/red-mobile-nodes": "^2.6.11",
+    "archiver": "^5.2.0",
+    "axios": "^0.21.1",
+    "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "fs-extra": "^9.0.1",
     "modclean": "^3.0.0-beta.1",
-    "nodejs-mobile-cordova": "git+https://github.com/okhiroyuki/nodejs-mobile-cordova.git"
-  },
-  "cordova": {
-    "plugins": {
-      "cordova-plugin-whitelist": {},
-      "nodejs-mobile-cordova": {}
-    },
-    "platforms": []
+    "multer": "^1.4.1",
+    "node-red": "^1.2.6",
+    "node-red-contrib-image-output": "^0.6.3",
+    "node-red-contrib-string": "^1.0.0",
+    "node-red-contrib-telegrambot": "^8.9.6",
+    "node-red-dashboard": "^2.26.1",
+    "node-red-node-base64": "^0.2.1",
+    "node-red-node-email": "^1.8.2",
+    "qs": "^6.9.4",
+    "uuid": "^8.3.2",
+    "ws": "^6.2.1"
   }
 }
