--- scripts/android/prepareNode.js
@@ -1,5 +1,0 @@
-module.exports = function(context) {
-    const execSync = require('child_process').execSync;
-    const result =  execSync('cd www/nodejs-project && bash prepare.sh').toString();
-    console.log(result);
-};
