--- www/NODEJS_MOBILE_BUILD_NATIVE_MODULES_VALUE.txt
@@ -1,1 +1,0 @@
-1
