--- www/nodejs-project/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "nodejs-mobile-sample-project",
-  "version": "0.1.0",
-  "description": "Node.js for Mobile Apps sample project",
-  "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "dependencies": {
-    "@red-mobile/red-mobile-nodes": "^2.6.11",
-    "axios": "^0.21.1",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "fs-extra": "^9.0.1",
-    "multer": "^1.4.1",
-    "node-red": "^1.2.6",
-    "node-red-contrib-image-output": "^0.6.3",
-    "node-red-contrib-string": "^1.0.0",
-    "node-red-contrib-telegrambot": "^8.9.6",
-    "node-red-dashboard": "^2.26.1",
-    "node-red-node-base64": "^0.2.1",
-    "node-red-node-email": "^1.8.2",
-    "qs": "^6.9.4",
-    "uuidv4": "^6.2.6",
-    "ws": "^6.2.1"
-  },
-  "author": "Janea Systems Inc.",
-  "license": "MIT"
-}
